requirements.txt
Django==4.2.11

# Configuração
python-decouple==3.8.0
dj-database-url>=2.1.0,<3
gunicorn>=22.0.0,<23.0.0
whitenoise==6.12.0  
python-dotenv==1.2.1

# Frontend (Django + Tailwind)
django-tailwind==3.8.0
django-browser-reload>=1.21.0,<2.0.0
django-unfold==0.89.0

# Imagens
Pillow==12.1.1

# Parsing / utilidades
beautifulsoup4>=4.12.2,<5

# Email
sendgrid>=6.11.0,<7